requirements.txt

fastapi==0.114.1
fastapi-cli==0.0.5
jsonify==0.5
psycopg2==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
SQLAlchemy==2.0.34
SQLAlchemy-Utils==0.41.2
uvicorn==0.30.6
pymysql==1.1.1
